backlogScreen抽象改代码前的备份

+++ app/src/main/java/com/sather/todo/ui/navigation/ToDoNavHost.kt
@@ -61,7 +61,7 @@ fun ToDoNavHost(
                     )
                 },
             ){
-                val baklogId=it.arguments?.getInt("backlogId")
+                val baklogId=it.arguments?.getLong("backlogId")
                 SingleBacklogScreen(
                     this@SharedTransitionLayout,
                     this@composable,
